Move ProjectPageViewModel to Kickstarter-Framework
