+++ Kickstarter-Framework/Sources/Kickstarter-Framework-iOSTests/Kickstarter-iOS/Features/ProjectPage/ViewModel/ProjectPageViewModelTests.swift
@@ -1,4 +1,7 @@
 import AVFoundation
+import Experimentation
+import GraphAPI
+@testable import Kickstarter_Framework
 @testable import KsApi
 @testable import KsApiTestHelpers
 @testable import Library
