+++ Kickstarter-Framework/Sources/Kickstarter-Framework/Kickstarter-iOS/Features/ProjectPage/ViewModel/ProjectPageViewModel.swift
@@ -1,28 +1,9 @@
 import Foundation
 import GraphAPI
 import KsApi
+import Library
 import Prelude
 import ReactiveSwift
-
-public protocol ProjectPageParam {
-  var param: Param { get }
-  var initialProject: (any ProjectPamphletMainCellConfiguration)? { get }
-}
-
-public struct ProjectPageParamBox: ProjectPageParam {
-  public let param: Param
-  public let initialProject: (any ProjectPamphletMainCellConfiguration)?
-
-  public init(param: Param, initialProject: (any ProjectPamphletMainCellConfiguration)?) {
-    self.param = param
-    self.initialProject = initialProject
-  }
-}
-
-extension Param: ProjectPageParam {
-  public var param: Param { self }
-  public var initialProject: (any ProjectPamphletMainCellConfiguration)? { nil }
-}
 
 public protocol ProjectPageViewModelInputs {
   /// Call when didSelectRowAt is called on a `ProjectFAQAskAQuestionCell`
